+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -3,6 +3,7 @@ import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { authController } from "~/modules/auth/auth.js";
 import { groupController } from "~/modules/groups/groups.js";
+import { permissionController } from "~/modules/permissions/permissions.js";
 import { projectGroupController } from "~/modules/project-groups/project-groups.js";
 import { projectController } from "~/modules/projects/projects.js";
 import { userController, userService } from "~/modules/users/users.js";
@@ -16,6 +17,7 @@ const apiV1 = new BaseServerApplicationApi(
 	"v1",
 	config,
 	...authController.routes,
+	...permissionController.routes,
 	...projectGroupController.routes,
 	...projectController.routes,
 	...userController.routes,