+++ apps/backend/src/modules/projects/project.repository.ts
@@ -27,8 +27,13 @@ class ProjectRepository implements Repository {
 		return ProjectEntity.initialize(user);
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const deletedRowsCount = await this.projectModel
+			.query()
+			.deleteById(id)
+			.execute();
+
+		return Boolean(deletedRowsCount);
 	}
 
 	public async find(id: number): Promise<null | ProjectEntity> {