+++ apps/backend/src/modules/users/user.repository.ts
@@ -3,10 +3,9 @@ import {
 	type PaginationResponseDto,
 	type Repository,
 } from "~/libs/types/types.js";
+import { type UserPatchRequestDto } from "~/modules/users/libs/types/types.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
-
-import { type UserPatchRequestDto } from "./libs/types/types.js";
 
 class UserRepository implements Repository {
 	private userModel: typeof UserModel;
@@ -31,15 +30,29 @@ class UserRepository implements Repository {
 		return UserEntity.initialize(user);
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const deletedRowsCount = await this.userModel
+			.query()
+			.patch({ deletedAt: new Date().toISOString() })
+			.where({ id })
+			.whereNull("deletedAt")
+			.execute();
+
+		return Boolean(deletedRowsCount);
 	}
 
 	public async find(id: number): Promise<null | UserEntity> {
 		const user = await this.userModel
 			.query()
 			.findById(id)
-			.returning("*")
+			.whereNull("deletedAt")
+			.withGraphFetched("groups.permissions")
+			.modifyGraph("groups", (builder) => {
+				builder.select("user_groups.id", "name");
+			})
+			.modifyGraph("groups.permissions", (builder) => {
+				builder.select("permissions.id", "name", "key");
+			})
 			.execute();
 
 		return user ? UserEntity.initialize(user) : null;
@@ -51,6 +64,7 @@ class UserRepository implements Repository {
 	}: PaginationQueryParameters): Promise<PaginationResponseDto<UserEntity>> {
 		const { results, total } = await this.userModel
 			.query()
+			.whereNull("deletedAt")
 			.page(page, pageSize);
 
 		return {
@@ -59,8 +73,26 @@ class UserRepository implements Repository {
 		};
 	}
 
-	public async findByEmail(email: string): Promise<null | UserEntity> {
-		const user = await this.userModel.query().findOne({ email });
+	public async findByEmail(
+		email: string,
+		hasDeleted = false,
+	): Promise<null | UserEntity> {
+		const query = this.userModel
+			.query()
+			.findOne({ email })
+			.withGraphFetched("groups.permissions")
+			.modifyGraph("user_groups.id", (builder) => {
+				builder.select("id", "name");
+			})
+			.modifyGraph("groups.permissions", (builder) => {
+				builder.select("permissions.id", "name", "key");
+			});
+
+		if (!hasDeleted) {
+			query.whereNull("deletedAt");
+		}
+
+		const user = await query.execute();
 
 		return user ? UserEntity.initialize(user) : null;
 	}