+++ apps/frontend/src/index.tsx
@@ -8,7 +8,7 @@ import {
 	StoreProvider,
 	ToastContainer,
 } from "~/libs/components/components.js";
-import { AppRoute } from "~/libs/enums/enums.js";
+import { AppRoute, PermissionKey } from "~/libs/enums/enums.js";
 import { store } from "~/libs/modules/store/store.js";
 import { AccessManagement } from "~/pages/access-management/access-management.jsx";
 import { Analytics } from "~/pages/analytics/analytics.jsx";
@@ -38,7 +38,9 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							},
 							{
 								element: (
-									<ProtectedRoute>
+									<ProtectedRoute
+										routePermissions={[PermissionKey.MANAGE_USER_ACCESS]}
+									>
 										<AccessManagement />
 									</ProtectedRoute>
 								),