+++ apps/frontend/src/libs/constants/navigation-items.constant.ts
@@ -1,4 +1,4 @@
-import { AppRoute } from "~/libs/enums/enums.js";
+import { AppRoute, PermissionKey } from "~/libs/enums/enums.js";
 import { type NavigationItem } from "~/libs/types/types.js";
 
 const SIDEBAR_ITEMS: NavigationItem[] = [
@@ -11,6 +11,7 @@ const SIDEBAR_ITEMS: NavigationItem[] = [
 		href: AppRoute.ACCESS_MANAGEMENT,
 		icon: "access",
 		label: "Access Management",
+		pagePermissions: [PermissionKey.MANAGE_USER_ACCESS],
 	},
 	{
 		href: AppRoute.ANALYTICS,