+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -1,7 +1,9 @@
 const NotificationMessage = {
+	PROFILE_UPDATE_SUCCESS: "Successfully updated profile information.",
 	PROJECT_CREATE_SUCCESS: "Project was successfully created",
+	PROJECT_DELETE_SUCCESS: "Project was successfully deleted.",
 	PROJECT_UPDATE_SUCCESS: "Project was successfully updated.",
-	SUCCESS_PROFILE_UPDATE: "Successfully updated profile information.",
+	USER_DELETE_SUCCESS: "User deleted successfully.",
 } as const;
 
 export { NotificationMessage };