+++ apps/frontend/src/libs/helpers/helpers.ts
@@ -1,3 +1,8 @@
 export { getShortWeekday } from "./get-short-weekday.helper.js";
 export { getValidClassNames } from "./get-valid-classes.helper.js";
-export { configureString, formatDate, initDebounce } from "@git-fit/shared";
+export {
+	checkHasPermission,
+	configureString,
+	formatDate,
+	initDebounce,
+} from "@git-fit/shared";