+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -13,6 +13,7 @@ type IconName =
 	| "rightArrow"
 	| "rightDoubleArrow"
 	| "search"
-	| "strikedEye";
+	| "strikedEye"
+	| "trashBin";
 
 export { type IconName };