+++ packages/shared/src/modules/users/libs/types/user-auth-response-dto.type.ts
@@ -1,6 +1,15 @@
 type UserAuthResponseDto = {
 	createdAt: string;
 	email: string;
+	groups: {
+		id: number;
+		name: string;
+		permissions: {
+			id: number;
+			key: string;
+			name: string;
+		}[];
+	}[];
 	id: number;
 	name: string;
 };